requirements.txt
pandas==2.0.0
numpy==1.24.0
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.0
yfinance==0.2.28
tweepy==4.14.0
schedule==1.2.0
python-dotenv==1.0.0
pyyaml==6.0
loguru==0.7.0
line-bot-sdk==3.5.0
ta==0.11.0
feedparser==6.0.10